Segundo cambio

+++ src/main/java/com/example/demoods/service/ComentarioServicioImpl.java
@@ -3,9 +3,11 @@ package com.example.demoods.service;
 import com.example.demoods.dto.ComentarioDTO;
 import com.example.demoods.entity.Comentario;
 import com.example.demoods.entity.Publicacion;
+import com.example.demoods.exeption.BlogAppException;
 import com.example.demoods.exeption.ResourceNotFoundException;
 import com.example.demoods.repository.IComentarioRepository;
 import com.example.demoods.repository.IPublicacionRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +22,8 @@ public class ComentarioServicioImpl implements IComentarioServicio{
     private IComentarioRepository iComentarioRepository;
     @Autowired
     private IPublicacionRepository iPublicacionRepository;
+    @Autowired
+    private ModelMapper modelMapper;
     @Override
     public ComentarioDTO guardarComentario(long publicacionId, ComentarioDTO comentarioDTO) {
         Comentario comentario=convertirComentario(comentarioDTO);
@@ -30,42 +34,66 @@ public class ComentarioServicioImpl implements IComentarioServicio{
     }
 
     @Override
-    public List<ComentarioDTO> obtenerComentarioPorPublicacionId(long publicacionId) {
+    public List<ComentarioDTO> obtenerComentarioPorPublicacionId(Long publicacionId) {
         List<Comentario> comentarioList = iComentarioRepository.findByPublicacionId(publicacionId);
         return comentarioList.stream().map(comentario -> convertirComentarioDTO(comentario)).collect(Collectors.toList());
     }
 
     @Override
-    public ComentarioDTO obtenerComentarioPorId(long comentarioId) {
-        Comentario comentario=iComentarioRepository.findById(comentarioId).orElseThrow(() ->new  ResourceNotFoundException("Comentario","id",comentarioId));
+    public ComentarioDTO obtenerComentarioPorId(Long publicacionId,Long comentarioId) {
+        Publicacion publicacion=iPublicacionRepository.findById(publicacionId)
+                .orElseThrow(()-> new ResourceNotFoundException("Publicacion","id",publicacionId));
+        Comentario comentario=iComentarioRepository.findById(comentarioId)
+                .orElseThrow(() ->new  ResourceNotFoundException("Comentario","id",comentarioId));
+        if(!comentario.getPublicacion().getId().equals(publicacion.getId())){
+            throw new BlogAppException(HttpStatus.BAD_REQUEST,"Comentario no encontrado");
+        }
+
         return convertirComentarioDTO(comentario);
+
     }
 
     @Override
-    public void eliminarComentarioPorId(long comentarioId) {
-        Comentario comentario=iComentarioRepository.findById(comentarioId).orElseThrow(() ->new  ResourceNotFoundException("Comentario","id",comentarioId));
+    public void eliminarComentarioPorId(Long publicacionId,Long comentarioId) {
+        Publicacion publicacion=iPublicacionRepository.findById(publicacionId)
+                .orElseThrow(()-> new ResourceNotFoundException("Publicacion","id",publicacionId));
+        Comentario comentario=iComentarioRepository.findById(comentarioId)
+                .orElseThrow(() ->new  ResourceNotFoundException("Comentario","id",comentarioId));
+        if(!comentario.getPublicacion().getId().equals(publicacion.getId())){
+            throw new BlogAppException(HttpStatus.BAD_REQUEST,"Comentario no encontrado");
+        }
         iComentarioRepository.delete(comentario);
 
     }
 
+    @Override
+    public ComentarioDTO actualizarComentario(Long publicacionId, ComentarioDTO cuerpoComentarioDTO,Long comentarioId) {
+        Publicacion publicacion=iPublicacionRepository.findById(publicacionId)
+                .orElseThrow(()-> new ResourceNotFoundException("Publicacion","id",publicacionId));
+        Comentario comentario=iComentarioRepository.findById(comentarioId)
+                .orElseThrow(() ->new  ResourceNotFoundException("Comentario","id",comentarioId));
+        if(!comentario.getPublicacion().getId().equals(publicacion.getId())){
+            throw new BlogAppException(HttpStatus.BAD_REQUEST,"Comentario no encontrado");
+        }
+        comentario.setNombre(cuerpoComentarioDTO.getNombre());
+        comentario.setEmail(cuerpoComentarioDTO.getEmail());
+        comentario.setCuerpo(cuerpoComentarioDTO.getCuerpo());
+
+        Comentario comentarioActualizado=iComentarioRepository.save(comentario);
+
+        return convertirComentarioDTO(comentarioActualizado);
+    }
+
     //Convertir a DTO
     private ComentarioDTO convertirComentarioDTO(Comentario comentario){
-        ComentarioDTO comentarioDTO=new ComentarioDTO();
-        comentarioDTO.setId(comentario.getId());
-        comentarioDTO.setNombre(comentario.getNombre());
-        comentarioDTO.setEmail(comentario.getEmail());
-        comentarioDTO.setCuerpo(comentario.getCuerpo());
+        ComentarioDTO comentarioDTO=modelMapper.map(comentario,ComentarioDTO.class);
         return comentarioDTO;
 
     }
 
     //Convertir a Entidad
     private Comentario convertirComentario(ComentarioDTO comentarioDTO){
-        Comentario comentario=new Comentario();
-        comentario.setId(comentarioDTO.getId());
-        comentario.setNombre(comentarioDTO.getNombre());
-        comentario.setEmail(comentarioDTO.getEmail());
-        comentario.setCuerpo(comentarioDTO.getCuerpo());
+        Comentario comentario=modelMapper.map(comentarioDTO,Comentario.class);
         return comentario;
     }
 }
